# docs-site/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tiao API",
    "version": "1.0.0",
    "description": "REST API and WebSocket protocol for Tiao, the open-source multiplayer board game platform."
  },
  "servers": [
    {
      "url": "http://localhost:5005/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/games": {
      "get": {
        "summary": "List the current player's games",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of games",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "games": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MultiplayerSnapshot"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new multiplayer game",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Game created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "snapshot": {
                      "$ref": "#/components/schemas/MultiplayerSnapshot"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/games/{gameId}/join": {
      "post": {
        "summary": "Join an existing game",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Joined the game",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "snapshot": {
                      "$ref": "#/components/schemas/MultiplayerSnapshot"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/games/{gameId}/access": {
      "post": {
        "summary": "Access a game (join or spectate)",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game accessed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "snapshot": {
                      "$ref": "#/components/schemas/MultiplayerSnapshot"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/games/{gameId}": {
      "get": {
        "summary": "Get a game snapshot",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "snapshot": {
                      "$ref": "#/components/schemas/MultiplayerSnapshot"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/matchmaking": {
      "post": {
        "summary": "Enter matchmaking queue",
        "tags": [
          "Matchmaking"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Entered matchmaking",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "matchmaking": {
                      "$ref": "#/components/schemas/MatchmakingState"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "summary": "Get current matchmaking status",
        "tags": [
          "Matchmaking"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Matchmaking state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "matchmaking": {
                      "$ref": "#/components/schemas/MatchmakingState"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Leave matchmaking queue",
        "tags": [
          "Matchmaking"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "Left matchmaking"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/{gameId}/test-finish": {
      "post": {
        "summary": "Force-finish a game (development only)",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "gameId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "winner"
                ],
                "properties": {
                  "winner": {
                    "type": "string",
                    "enum": [
                      "white",
                      "black"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game finished",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not allowed in production"
          },
          "500": {
            "description": "Failed to finish game"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MultiplayerSnapshot": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string"
          },
          "roomType": {
            "type": "string",
            "enum": [
              "direct",
              "matchmaking"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "waiting",
              "active",
              "finished"
            ]
          },
          "state": {
            "type": "object"
          },
          "players": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "seats": {
            "type": "object"
          },
          "rematch": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "MatchmakingState": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "searching",
              "matched"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tiao.session"
      }
    }
  }
}